change split genomes from output to param
The `directory` directive touches an empty file within the directory.
This causes WIsH to iterate over that file and messes its output. It
contains an empty string as contig and contains an empty phage entry
with a random taxonomy assignment. By making it a parameter for the
rules the dir only contains fasta files.